backend/requirements.txt
flask
flask-cors
openai-whisper
llama-cpp-python
edge-tts
asyncio
